+++ test/comment-list-one.test.js
@@ -49,6 +49,7 @@ describe('Test One Comment', ()=>{
                 push: "https://i5daxohp.push.lncld.net",
                 stats: "https://i5daxohp.stats.lncld.net"
               }}
+              CommentClass={"Comment"}
       >
         <div className="App">
           <header className="App-header">