+++ test/fetch-more-comment-list.test.js
@@ -71,6 +71,7 @@ describe('Test Multiple Comments', ()=> {
                 push: "https://i5daxohp.push.lncld.net",
                 stats: "https://i5daxohp.stats.lncld.net"
               }}
+              CommentClass={"Comment"}
       >
         <div className="App">
           <header className="App-header">