+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * `API`请求替换`leancloud-sdk`，显著减少包大小(parsed size from `479.2kb` to `331.92kb`)
 * 自定义组件UI替换`material-ui`，显著减少包大小(parsed size from `331.92kb` to `216.48kb`)
+* 增加自定义`leancloud`上Class名称，默认为`"Comment_demo"`
 
 ##### 0.5.0 (2020-3-13)